+++ src/main/java/com/teamabode/verdance/core/integration/farmersdelight/FDIntegration.java
@@ -1,7 +1,9 @@
 package com.teamabode.verdance.core.integration.farmersdelight;
 
 import com.teamabode.verdance.core.integration.CompatUtils;
+import com.teamabode.verdance.core.integration.farmersdelight.registry.FDCompatBlockEntityTypes;
 import com.teamabode.verdance.core.integration.farmersdelight.registry.FDCompatBlocks;
+import com.teamabode.verdance.core.integration.farmersdelight.registry.FDCompatSoundEvents;
 import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
 import net.fabricmc.loader.api.ModContainer;
 import net.minecraft.item.ItemGroup;
@@ -19,6 +21,9 @@ public class FDIntegration {
 
     public static void register(ModContainer container) {
         FDCompatBlocks.register();
+        FDCompatBlockEntityTypes.register();
+        FDCompatSoundEvents.register();
+
         ItemGroupEvents.modifyEntriesEvent(TAB_FARMERS_DELIGHT).register(entries -> {
             entries.addAfter(ModItems.ONION_CRATE.get(), FDCompatBlocks.MULBERRY_CRATE);
             entries.addAfter(ModBlocks.CHERRY_CABINET.get(), FDCompatBlocks.MULBERRY_CABINET);