+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## Tweaks
+* Cantaloupe Juice only shows the extinguish effects when on fire.
+
 ## Fixes
+* Fixed an issue where potted variants of plant blocks had registered block items.
+* Fixed an issue where Cantaloupe Juice consumed 2 bottles.
 * Fixed `verdance:cantaloupe` lacking the `minecraft:sword_efficient` block tag.